Add the route context to the request map
This provides access to the route prefix as a companion to the existing
:compojure/route data allowing the full route to be reconstructed.
Since the context macro only accepts a path the new :compojure/context
value simply contains the route path instead of the method+route pair
present in :compojure/route.

This data was introduced as a new field in the request path, rather than
updating :compojure/route to include the context for the sake of
backwards compatibility.