Add a full-route key to the request map
This provides access to the full route that contains the common prefix.
The full-route info is added to the :compojure/full-route key in the
request map.

We use :compojure/route to get the route info and add it to the
attributes of our metrics and traces, but when compojure.core/context
is used, we are not able to get the parameters that are not
instantiated. This change adds :compojure/full-route key that serves
similarly as the existing :compojure/route does - the only difference
is that the new key has a common prefix.

Co-authored-by: Liam Chen <liamchzh@gmail.com>
Co-authored-by: Claire Alvis <claire.alvis@gmail.com>